moved backup function to app file and made max backup variable
